user-service/requirements.txt
# User Service Dependencies

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Password hashing
passlib[bcrypt]==1.7.4
pwdlib[argon2]

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Environment
python-dotenv==1.0.0

# Monitoring
prometheus-fastapi-instrumentator

# OpenTelemetry Tracing  
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger-thrift
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-httpx
structlog